requirements.txt
streamlit==1.31.0
opencv-python-headless==4.8.1.78
tensorflow==2.15.0
numpy==1.24.3
Pillow==10.0.0